pyproject.toml
[project]
name = "terminal-dashboard"
version = "0.5.0"
description = "Terminal Dashboard for Odoo Data Visualization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.0",
    "SQLAlchemy==2.0.41",
    "aiosqlite==0.21.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.5.0",
    "email-validator>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "httpx>=0.25.0",
]
